requirements.txt
requests
python-telegram-bot
IMDbPY